addning school CRUD

+++ server/src/school/school.service.ts
@@ -1,55 +1,82 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 
 @Injectable()
 export class SchoolService {
-  constructor(private prisma: PrismaService) {}
+  constructor(private readonly prisma: PrismaService) {}
 
-  //create a school
-  async createSchool(data: { name: string }) {
-    const school = await this.prisma.school.create({
+  async registerSchools(data: { name: string }) {
+    const registerSchool = await this.prisma.school.create({
       data,
     });
     return {
-      message: `${school.name} waa la abuuray`,
-      school: school,
+      message: `School-ka "${registerSchool.name}" waa la abuuray`,
+      id: registerSchool.id,
+      school: registerSchool.name,
+      created: registerSchool.createdAtd,
     };
   }
 
-  // update a school
+  // update
 
-  async updateSchool(id: string, name: string) {
+  async updateSchool(id: string, newName?: string) {
+    const school = await this.prisma.school.findUnique({
+      where: { id },
+    });
+
+    if (!school) {
+      throw new NotFoundException('xogta lama helin'); // Not found
+    }
+
+    if (!newName) {
+      throw new BadRequestException('Magaca cusub waa lama huraan'); // Name missing
+    }
+
     const updatedSchool = await this.prisma.school.update({
       where: { id },
-      data: {
-        name,
-      },
+      data: { name: newName },
     });
+
     return {
-      message: 'Waad ku guulaysatay inaad wax ka badasho',
-      school_data: updatedSchool,
+      message: 'Xogta waa la cusboonaysiiyay',
+      new_name: updatedSchool.name,
     };
   }
 
-  // only admins for get all the school data
-  async getAllSchools() {
-    const schols = await this.prisma.school.findMany({});
+  // get all schools only admins and debuging purpose
 
-    return schols;
+  async readSchoolData() {
+    const schoolData = await this.prisma.school.findMany();
+
+    if (!schoolData) {
+      throw new NotFoundException('Xog maad diwaangalin');
+    }
+    return schoolData;
   }
 
-  async deleteSchool(id: string) {
-    const deletedSchool = await this.prisma.school.delete({
-      where: { id },
-    });
+  // delete statement
 
-    if (!deletedSchool)
+  async deletedSchool(id: string) {
+    try {
+      const existingSchool = await this.prisma.school.findUnique({
+        where: { id: id },
+      });
+
+      if (!existingSchool) {
+        throw new NotFoundException('Xog hore ayaa loo tirtiray');
+      }
+
+      return this.prisma.school.delete({
+        where: { id: id },
+      });
+    } catch (error: any) {
       return {
-        message: 'Hore ayaa loo tirtiray',
+        message: `Interal error ${error}`,
       };
-    else
-      return {
-        message: `waad ku guulaysatay inaad saarto (${deletedSchool.name})`,
-      };
+    }
   }
 }
